Title: 바이러스 검사, Time: 462ms, Memory: 74MB, Status: Passed - CodeTree

+++ 250311/바이러스 검사/virus-detector.java
@@ -19,7 +19,7 @@ public class Main {
         int leader = Integer.parseInt(st.nextToken());
         int member = Integer.parseInt(st.nextToken());
 
-        int memberCnt = 0;
+        long memberCnt = 0;
         for (int cust : custs) {
             cust -= leader;
             if (cust <= 0) continue;
